Refactor file_with_binary_template test
This commit modifies the file_with_binary_template test to use the
create_remote_site_pp helper to create the environment path.

The curl call to the pdb query interface has been modified to bind
to the database host name.

The create_remote_site_pp helper is also modified to create a moduledir
as part of the environment creation.